chore: Refactor the dcmparser module #2 BREAKING CHANGE: imports to dcmparser and dcmparser_util must be updated
