+++ dcmpipe_cli/src/app/mod.rs
@@ -5,7 +5,7 @@ use std::path::Path;
 use anyhow::{anyhow, Result};
 use dcmpipe_dict::dict::stdlookup::STANDARD_DICOM_DICTIONARY;
 use dcmpipe_lib::core::dcmelement::DicomElement;
-use dcmpipe_lib::core::dcmparser::{ParseError, Parser, ParserBuilder};
+use dcmpipe_lib::core::parser::{error::ParseError, parser::Parser, builder::ParserBuilder};
 
 pub(crate) mod archiveapp;
 pub(crate) mod editapp;
