+++ dcmpipe_cli/src/app/printapp.rs
@@ -8,8 +8,8 @@ use dcmpipe_dict::dict::stdlookup::STANDARD_DICOM_DICTIONARY;
 use dcmpipe_dict::dict::tags;
 use dcmpipe_lib::core::dcmelement::{DicomElement, RawValue};
 use dcmpipe_lib::core::dcmobject::{DicomNode, DicomObject, DicomRoot};
-use dcmpipe_lib::core::dcmparser::Parser;
-use dcmpipe_lib::core::dcmparser_util::parse_into_object;
+use dcmpipe_lib::core::parser::parser::Parser;
+use dcmpipe_lib::core::parser::util::parse_into_object;
 use dcmpipe_lib::core::dcmsqelem::SequenceElement;
 use dcmpipe_lib::defn::dcmdict::DicomDictionary;
 use dcmpipe_lib::defn::tag::Tag;
