+++ dcmpipe_lib/src/core/mod.rs
@@ -1,7 +1,14 @@
+pub use parser::parser::Parser;
+pub use parser::error::ParseError;
+pub use parser::parser::Result;
+pub use parser::parser::DICOM_PREFIX;
+pub use parser::parser::DICOM_PREFIX_LENGTH;
+pub use parser::parser::FILE_PREAMBLE_LENGTH;
+pub use parser::builder::ParserBuilder;
+
+pub mod parser;
 pub mod charset;
 pub mod dcmelement;
 pub mod dcmobject;
-pub mod dcmparser;
-pub mod dcmparser_util;
 pub mod dcmsqelem;
 pub mod tagstop;
