+++ dcmpipe_lib/src/core/parser/util.rs
@@ -3,7 +3,8 @@ use std::io::{ErrorKind, Read};
 
 use crate::core::dcmelement::DicomElement;
 use crate::core::dcmobject::{DicomObject, DicomRoot};
-use crate::core::dcmparser::{ParseError, Parser, Result};
+use crate::core::parser::error::{ParseError};
+use super::parser::{Parser, Result};
 use crate::defn::constants::tags;
 use crate::defn::vl;
 use crate::defn::vl::ValueLength;
@@ -19,8 +20,10 @@ pub(crate) fn is_non_standard_seq(tag: u32, vr: VRRef, vl: ValueLength) -> bool 
         && vl == ValueLength::UndefinedLength
 }
 
+
 /// This is a variation of `Read::read_exact` however if zero bytes are read instead of returning
 /// an error with `ErrorKind::UnexpectedEof` it will return an error with `ParseError::ExpectedEOF`.
+#[allow(dead_code)]
 fn read_exact_expect_eof(dataset: &mut impl Read, mut buf: &mut [u8]) -> Result<()> {
     let mut bytes_read: usize = 0;
     while !buf.is_empty() {
@@ -52,6 +55,7 @@ fn read_exact_expect_eof(dataset: &mut impl Read, mut buf: &mut [u8]) -> Result<
 }
 
 /// Reads a tag attribute from a given dataset
+#[allow(dead_code)]
 pub(crate) fn read_tag_from_dataset(dataset: &mut impl Read, big_endian: bool) -> Result<u32> {
     let mut buf: [u8; 2] = [0; 2];
 
@@ -131,6 +135,7 @@ pub(crate) fn read_value_length_from_dataset(
 /// Parses elements to build a `DicomObject` to represent the parsed dataset as an in-memory tree.
 /// Returns `None` if the parser's first element fails to parse properly, assumed to be a non-DICOM
 /// dataset. Any errors after a successful first element being parsed are returned as `Result::Err`.
+#[allow(dead_code)]
 pub fn parse_into_object<'dict, DatasetType: Read>(
     parser: &mut Parser<'dict, DatasetType>,
 ) -> Result<Option<DicomRoot<'dict>>> {
@@ -168,6 +173,7 @@ pub fn parse_into_object<'dict, DatasetType: Read>(
 /// `child_nodes` The map of nodes which elements should be parsed into
 /// `item_nodes` The list of nodes which item elements should be parsed into
 /// `is_first_level` Whether the root level is being parsed, or within child nodes
+#[allow(dead_code)]
 fn parse_into_object_recurse<DatasetType: Read>(
     parser: &mut Parser<'_, DatasetType>,
     child_nodes: &mut BTreeMap<u32, DicomObject>,
