+++ dcmpipe_tests/src/lib.rs
@@ -6,10 +6,11 @@ use walkdir::WalkDir;
 
 use dcmpipe_dict::dict::stdlookup::STANDARD_DICOM_DICTIONARY;
 use dcmpipe_lib::core::dcmobject::{DicomNode, DicomObject, DicomRoot};
-use dcmpipe_lib::core::dcmparser::{
-    Parser, ParserBuilder, Result, DICOM_PREFIX, DICOM_PREFIX_LENGTH, FILE_PREAMBLE_LENGTH,
+use dcmpipe_lib::core::parser::{
+    parser::Parser, builder::ParserBuilder, parser::Result, parser::DICOM_PREFIX, parser::DICOM_PREFIX_LENGTH, parser::FILE_PREAMBLE_LENGTH,
 };
-use dcmpipe_lib::core::dcmparser_util::parse_into_object;
+
+use dcmpipe_lib::core::parser::util::parse_into_object;
 use dcmpipe_lib::core::tagstop::TagStop;
 use dcmpipe_lib::defn::tag::Tag;
